Overall fixes to the video playback

+++ vml-streamer/dpg_callback.py
@@ -2,32 +2,26 @@ import sys, subprocess
 import dearpygui.dearpygui as dpg
 import numpy as np
 import stream_types as st
-import platform
+import platform, time
 
 # recreates raw texture on registry
 def recreate_raw_texture(width=320, height=240):
 	
-	with dpg.mutex():
-	
-		# del image
-		if dpg.does_alias_exist('video_image'): 
-			dpg.delete_item('video_image')
-			try: dpg.remove_alias('video_image')
-			except: pass
+	# del texture
+	if dpg.does_alias_exist('cv_frame'): dpg.delete_item('cv_frame')	
+	if dpg.does_alias_exist('cv_frame'): dpg.remove_alias('cv_frame')		
 
-		# del texture
-		if dpg.does_alias_exist('cv_frame'): 
-			dpg.delete_item('cv_frame')
-			try: dpg.remove_alias('cv_frame')
-			except: pass
+	# del image
+	if dpg.does_alias_exist('video_image'): dpg.delete_item('video_image')
+	if dpg.does_alias_exist('video_image'): dpg.remove_alias('video_image')
 
-		# recreate texture
-		texture_data = np.zeros(shape=(width, height, 3))
-		dpg.add_raw_texture(width, height, texture_data, format=dpg.mvFormat_Float_rgb, tag='cv_frame', parent='treg')
-
-		#recreate image
-		if dpg.does_alias_exist('video_image_parent'): 
-			dpg.add_image(texture_tag='cv_frame', tag='video_image', parent='video_image_parent')
+	# recreate texture
+	texture_data = np.zeros(shape=(width, height, 3))
+	dpg.add_raw_texture(width, height, texture_data, format=dpg.mvFormat_Float_rgb, tag='cv_frame', parent='treg')
+			
+	# recreate image	
+	if dpg.does_alias_exist('video_image_parent'): 
+		dpg.add_image(texture_tag='cv_frame', tag='video_image', parent='video_image_parent', width=width, height=height)
 
 # remap value between range
 def change_range(unscaled, from_min, from_max, to_min, to_max):
@@ -46,8 +40,11 @@ def resize_img(sender, app_data, user_data):
 
 # resize viewport to specific values
 def resize_viewport(w=None, h=None):
+	pass
+	'''
 	if w: dpg.configure_viewport(0, width=w+18)
 	if h: dpg.configure_viewport(0, height=h)
+	'''
 
 # set viewport "always on top" mode
 def always_on_top(sender):
